use policy for both clients
